Update Reserve and Add Exception Update Review Update saveUpdate (need to solve runtime 404)

+++ src/main/java/com/one/eatmotion/service/TextReviewService.java
@@ -1,5 +1,6 @@
 package com.one.eatmotion.service;
 
+import com.one.eatmotion.entity.User;
 import com.one.eatmotion.entity.review.TextReview;
 import com.one.eatmotion.repository.TextReviewRepository;
 import lombok.RequiredArgsConstructor;
@@ -7,6 +8,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -15,6 +17,8 @@ public class TextReviewService {
 
   private final TextReviewRepository textReviewRepository;
   private final SentimentService sentimentService;
+  private final UserService userService;
+  private final ShopService shopService;
 
   /** onetoone으로 Receipt랑 Face 관계 맺고 가져오는 방법 */
   public List<TextReview> findAllByShopId(Long shopId) {
@@ -22,22 +26,35 @@ public class TextReviewService {
   }
 
   @Transactional
-  public TextReview saveTextReview(TextReview textReview) {
-    textReviewRepository.save(textReview);
-    textReview.setGrade(Double.valueOf(sentimentService.sentiment(textReview.getContent())));
-    return textReview;
-  }
-
-  @Transactional
-  public TextReview updateTextReview(Long id, String content) {
-    TextReview textReview = textReviewRepository.getById(id);
+  public TextReview saveTextReview(String content, Long shopId) {
+    User user = userService.getAuthedUser();
+    TextReview textReview = new TextReview();
     textReview.setContent(content);
-    textReview.setGrade(Double.valueOf(sentimentService.sentiment(content)));
+    textReview.setUser(user);
+    textReview.setShop(shopService.findById(shopId));
+    textReview.setGrade(sentimentService.sentiment(textReview.getContent()));
+    return textReviewRepository.save(textReview);
+  }
+
+  @Transactional
+  public TextReview updateTextReview(Long id, String content) throws Exception {
+    TextReview textReview = textReviewRepository.getById(id);
+    User user = userService.getAuthedUser();
+    if (!Objects.equals(user.getId(), textReview.getUser().getId())) {
+      throw new Exception();
+    }
+    textReview.setContent(content);
+    textReview.setGrade(sentimentService.sentiment(content));
     return textReview;
   }
 
   @Transactional
-  public void deleteById(Long id) {
+  public void deleteById(Long id) throws Exception {
+    TextReview textReview = textReviewRepository.getById(id);
+    User user = userService.getAuthedUser();
+    if (!Objects.equals(user.getId(), textReview.getUser().getId())) {
+      throw new Exception();
+    }
     textReviewRepository.deleteById(id);
   }
 }
